requirements.txt
pillow
requests
transformers
accelerate
PyQt6
flask
tdqm
python-dotenv
--find-links https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio